requirements.txt
# =============================================================================
# INTELLIGENT INVOICE HEADER EXTRACTION SYSTEM
# Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# CORE DEPENDENCIES
# -----------------------------------------------------------------------------
# Configuration management
PyYAML>=6.0.1

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# -----------------------------------------------------------------------------
# PDF PROCESSING
# -----------------------------------------------------------------------------
# PDF to image conversion
pdf2image>=1.16.3
PyMuPDF>=1.23.0  # Alternative: faster PDF handling

# PDF text extraction (for digital PDFs)
pdfplumber>=0.10.0

# -----------------------------------------------------------------------------
# OCR ENGINES
# -----------------------------------------------------------------------------
# Primary OCR engine
pytesseract>=0.3.10

# Alternative OCR engines (optional)
# easyocr>=1.7.0  # Uncomment if needed
# paddleocr>=2.7.0  # Uncomment if needed

# -----------------------------------------------------------------------------
# MACHINE LEARNING / TRANSFORMERS
# -----------------------------------------------------------------------------
# Hugging Face Transformers
transformers>=4.35.0
torch>=2.1.0

# Layout-aware document understanding
# LayoutLMv3 dependencies
timm>=0.9.0
detectron2  # May require separate installation on Windows

# -----------------------------------------------------------------------------
# OUTPUT HANDLING
# -----------------------------------------------------------------------------
# Excel file generation
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Database
sqlalchemy>=2.0.0

# -----------------------------------------------------------------------------
# EVALUATION & REPORTING
# -----------------------------------------------------------------------------
# Metrics computation
scikit-learn>=1.3.0

# Progress bars
tqdm>=4.66.0

# -----------------------------------------------------------------------------
# UTILITIES
# -----------------------------------------------------------------------------
# Colored console output
colorama>=0.4.6

# Date parsing
python-dateutil>=2.8.2

# Type hints
typing-extensions>=4.8.0

# -----------------------------------------------------------------------------
# DEVELOPMENT DEPENDENCIES (optional)
# -----------------------------------------------------------------------------
# Testing
# pytest>=7.4.0
# pytest-cov>=4.1.0

# Code formatting
# black>=23.0.0
# isort>=5.12.0

# Linting
# flake8>=6.1.0
# pylint>=3.0.0

# Documentation
# sphinx>=7.0.0
# sphinx-rtd-theme>=1.3.0
